backend/requirements.txt
groq
pdfplumber
python-dotenv
rich
